+++ lib/terminal_decorate.dart
@@ -1,1 +1,4 @@
-export 'src/standard.dart';
+export 'src/decorate_color.dart';
+export 'src/decorate_font.dart';
+export 'src/escape.dart';
+export 'src/wrap.dart';