+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 ## 0.1.0
 
 - Support foreground and background decorations.
+- Support combined complex foreground and background decorations.
+- Support combined complex font style decorations.
+- Tests to validate nested decorations.
 
 ## 0.0.0